[Helm] Add flags needed for migration from distributed (#7331)
**What this PR does / why we need it**:
Adds flags to the helm charts that are needed when doing a zero-downtime
migration from a `loki-distributed` helm chart deployment. The flags
allow for an additional memberlist join member to be specified. This
allows the new `grafana/loki` cluster to join the old
`grafana/loki-distributed` cluster, forming one big cluster. The
ingesters in the `grafana/loki-distributed` cluster can then be taken
down 1-by-1, allowing the whole `grafana/loki-distributed` cluster to be
deleted once all the ingesters have scaled down.

**Which issue(s) this PR fixes**:
Fixes #7330

Co-authored-by: Karsten Jeschkies <k@jeschkies.xyz>