Index gw lock bugfix (#7497)
While I'm still in favor of a longer refactor to address the complexity
described in https://github.com/grafana/loki/pull/7490, I'm opening this
PR as a shorter, more immediate alternative that should fix the
deadlocks. This also reduces the number of total lock calls in many
cases as we lock to act over entire groups rather than individual items.