Loki: Remove the bypass for "limited" queries  (#7510)
**What this PR does / why we need it**:

Limited queries are queries which don't have a filter expression. 

Now all log queries will be handled by the LogFilterTripperware which
will result in them being split by time (they previously were not)

Signed-off-by: Edward Welch <edward.welch@grafana.com>


**Which issue(s) this PR fixes**:

We've found that very large timeframe `limited` queries will be sent to
queriers and then ingesters and can stall out the read path because of
the large time ranges.

Splitting these by time avoids this problem by keeping any subquery
limited in length to the `split_by_interval`.

It's important to note that this will likely increase some extra work
done by limited queries as more work for these will be parallelized and
can result in extra data being processed. This will be the same as how
filter queries are handled now, so it will be no worse than that.

In fact this was an optimization based on the premise that it's
advantageous not to split/shard limited queries but we are seeing this
not be the case when limited queries are run over very large time ranges
matching streams with huge volumes and combined with parsers like `|
json`

**Special notes for your reviewer**:

**Checklist**
- [ ] Reviewed the `CONTRIBUTING.md` guide
- [ ] Documentation added
- [x] Tests updated
- [x] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`

Signed-off-by: Edward Welch <edward.welch@grafana.com>