fix logs deletion integration test to properly validate deleted data (#7751)
**What this PR does / why we need it**:
Currently, the integration test for deletion does not validate whether
the delete request properly deleted the requested data or not. It just
currently checks if the delete request got processed. This was due to
the singleton `boltdb-shipper` client causing problems in flushing
chunks.

I have used `ResetBoltDBIndexClientWithShipper` to reset the
boltdb-shipper client to get the chunk flush from the ingester working.
The test now does the checks to ensure that the right data got deleted
by the compactor. This should help us catch any bugs earlier for any
future changes in deletion code.

**Checklist**
- [x] Tests updated