[Helm] Remove unused value and set default values explicitly (#7576)
**What this PR does / why we need it**:
`monitoring.alerts` are not used in the Loki Helm chart. This removes
it.
Furthermore, some values are referenced in the templates, but not used
in the values file. This adds the default values.

**Which issue(s) this PR fixes**:
Fixes #7575

**Special notes for your reviewer**:

https://github.com/AndersBennedsgaard/loki/blob/main/production/helm/loki/templates/single-binary/pdb.yaml#L15
should probably also be updated since `podDisruptionBudget` is not
referenced in the values file. However, I am unsure what default values
should be used.

https://github.com/AndersBennedsgaard/loki/blob/main/production/helm/loki/templates/networkpolicy.yaml#L124
should probably also be updated since the `loki.rulerSelectorLabels` is
not a valid helper function. However, I am unsure which selector labels
to use. Especially since it seems like `ruler` is not used. Perhaps any
reference to `ruler` should be removed?

**Checklist**
- [x] Reviewed the `CONTRIBUTING.md` guide
- [x] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`